requirements.txt
fastapi
uvicorn
pymongo
motor
aiofiles
pandas
python-multipart
celery
redis
pillow
python-dotenv
requests